Added semester association when adding new program + generalized to concern
