+++ app/models/course.rb
@@ -2,10 +2,7 @@ class Course < ApplicationRecord
   include AASM
   include SystemVersioning
   include ApplicationVersioning
-  include VersioningHelper
-
-  attr_accessor :semester_season, :semester_year
-  before_create :add_semester_data
+  include SemesterDataAdder
 
   has_many :comments, as: :commentable
 
@@ -106,13 +103,6 @@ class Course < ApplicationRecord
     data['programs'] = programs
     data = data.as_json
     data
-  end
-
-  private
-
-  def add_semester_data
-    self.valid_end = get_valid_end_from_season_and_year(semester_season, semester_year.to_i)
-    self.valid_start = get_valid_start_from_valid_end(self.valid_end)
   end
 end
