+++ app/models/program.rb
@@ -1,6 +1,7 @@
 class Program < ApplicationRecord
   include SystemVersioning
   include ApplicationVersioning
+  include SemesterDataAdder
 
   has_many :course_programs, dependent: :destroy, :foreign_key => [:program_id, :program_valid_end]
   has_many :courses, through: :course_programs
