+++ app/views/programs/edit.html.erb
@@ -1,6 +1,6 @@
 <h1>Editing Program</h1>
 
-<%= render 'form', program: @program %>
+<%= render 'form', program: @program, is_create: false %>
 
 <%= link_to 'Show', @program %> |
 <%= link_to 'Back', programs_path %>
