+++ app/views/programs/new.html.erb
@@ -1,5 +1,5 @@
 <h1>New Program</h1>
 
-<%= render 'form', program: @program %>
+<%= render 'form', program: @program, is_create: true %>
 
 <%= link_to 'Back', programs_path %>
